chore: add eth example
