--- examples/FullSurface/configuration.yaml
@@ -1,60 +1,0 @@
-mysql:
-    host: localhost
-    user: root
-    password: password
-    database: DeribitOrderBook
-    use_bathes_to_record: False
-
-    reconnect_max_attempts: 5
-    reconnect_wait_time: 1
-
-hdf5:
-    hdf5_database_directory: "HDF_storage_24Jul"
-
-record_system:
-    use_batches_to_record: True
-    number_of_tmp_tables: 5
-    size_of_tmp_batch_table: 5
-    instrumentNameToIdMapFile: "InstrumentNameToIdMap.json"
-
-    clean_database_at_startup: False
-
-user_data:
-    test_net:
-        client_id: "W6-2Gwvq"
-        client_secret: "W9VQRlL7bIdc59DK2s7YrhqHTvw8k2U86nq_Tedsvfc"
-    production:
-        client_id: None
-        client_secret: None
-
-externalModules:
-    add_order_manager: False
-    add_instrument_manager: False
-
-
-orderBookScrapper:
-#    scrapper_body: [OrderBook, Trades, OwnOrderChange, Portfolio]
-    scrapper_body: [Trades, OrderBook]
-    depth: 10
-    test_net: False
-    currency: BTC
-    enable_traceback: False # False - default
-    enable_database_record: True
-    clean_database: False
-    hearth_beat_time: 60
-    group_in_limited_order_book: none
-    raise_error_at_synthetic: False # False - default
-    logger_level: INFO # WARN | INFO | ERROR
-
-    select_all_order_book: True
-    only_api_orders_processing: True
-
-    database_daemon: mysql
-
-    add_extra_instruments: ["BTC-PERPETUAL"]
-#     add_extra_instruments: ['BTC-5MAY23-28000-C', 'BTC-5MAY23-28000-P', 'BTC-5MAY23-30000-C', 'BTC-5MAY23-30500-P', 'BTC-5MAY23-50000-C', 'BTC-5MAY23-50000-P', 'BTC-5MAY23-51000-C', 'BTC-5MAY23-51000-P', 'BTC-5MAY23', 'BTC-PERPETUAL']
-#     add_extra_instruments: ['BTC-30JUN23-20000-C', 'BTC-30JUN23-20000-P', 'BTC-30JUN23-22000-C', 'BTC-30JUN23-22000-P', 'BTC-30JUN23-24000-C', 'BTC-30JUN23']
-    # Need to be true if you need to select maturities by configuration file.
-    use_configuration_to_select_maturities: False
-    maturities_configuration_path: "Maturities_configuration.yaml"
-
